+++ packages/core/src/types/index.ts
@@ -28,6 +28,5 @@ export {
   Emitted,
   RemoveResponseHandler,
   SideEffect,
-  Cache,
   Navigation
 } from "./curi";